+++ lib/src/engine/listeners/change_listener.dart
@@ -1,0 +1,5 @@
+abstract class ChangeListener {
+  void startListening();
+  void stopListening();
+  void dispose();
+}